Create highly modular architecture with focused single-responsibility modules
• Extract core copying logic to core.lua (81 lines)
• Extract commands and keybindings to commands.lua (50 lines)
• Simplify init.lua to ultra-clean 24-line orchestrator
• Update debug.lua to use new core module
• Maintain all functionality with better separation of concerns

Module responsibilities:
- init.lua: Setup orchestration and function exports
- core.lua: Core copying logic and utilities
- commands.lua: Command and keybinding setup
- explorers.lua: File explorer detection (167 lines)
- github.lua: GitHub permalink functionality (196 lines)
- debug.lua: Debug utilities and testing (80 lines)

Benefits: easier maintenance, better testability, clear separation of concerns
